refactor

+++ src/main/java/com/egecoskun/finalproject/model/DTO/ApplicantDTO.java
@@ -2,12 +2,11 @@ package com.egecoskun.finalproject.model.DTO;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 
 @Data
 public class ApplicantDTO {
 
-    private BigInteger identificationNumber;
+    private Long identificationNumber;
     private String firstName;
     private String lastName;
     private double monthlyIncome;
